tambahan konfirmasi delete

+++ src/main/java/view/MahasiswaTableView.java
@@ -200,15 +200,25 @@ public class MahasiswaTableView extends javax.swing.JFrame {
         // TODO add your handling code here:
         int barisTerpilih = tbMahasiswa.getSelectedRow();
         
-        if(barisTerpilih >= 0){
-            String npm = tbMahasiswa.getValueAt(barisTerpilih, 0).toString();
-            mahasiswa = new Mahasiswa();
-            mahasiswa.setNpm(npm);
-            mahasiswa.delete();
-            refresh();
-        }else{
-            JOptionPane.showMessageDialog(null, "Pilih dulu data");
+        int pilihan = JOptionPane.showConfirmDialog(null, 
+                "Data akan dihapus, lanjutkan?",
+                "Konfirmasi",
+                JOptionPane.YES_NO_OPTION);
+        
+        if(pilihan == 0){
+            if(barisTerpilih >= 0){
+                String npm = tbMahasiswa.getValueAt(barisTerpilih, 0).toString();
+                mahasiswa = new Mahasiswa();
+                mahasiswa.setNpm(npm);
+                mahasiswa.delete();
+                refresh();
+            }else{
+                JOptionPane.showMessageDialog(null, "Pilih dulu data");
+            }
         }
+        
+       
+
     }//GEN-LAST:event_btHapusActionPerformed
 
     private void btTutupActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btTutupActionPerformed
